+++ CHANGELOG.rst
@@ -51,6 +51,7 @@ Bugfixes
   lead to a frozen GUI.
 - Fixed an issue with plugin paths during unittests when production and 
   development versions are installed on the same machine.
+- Fixed an issue with consistency signals in PluginInWorkflowBox widget.
 
 
 v23.10.20